+++ src/resqml2_2/StreamlinesRepresentation.h
@@ -141,7 +141,7 @@ namespace RESQML2_2_NS
 		 * 										the repository default HDF proxy will be used.
 		 */
 		void setWellboreInformation(uint32_t const* injectorPerLine, uint32_t const* producerPerLine, const std::vector<RESQML2_NS::WellboreTrajectoryRepresentation*> & wellboreTrajectories,
-			uint32_t nullValue = (std::numeric_limits<uint64_t>::max)(), EML2_NS::AbstractHdfProxy* hdfProxy = nullptr) final;
+			uint32_t nullValue = (std::numeric_limits<uint32_t>::max)(), EML2_NS::AbstractHdfProxy* hdfProxy = nullptr) final;
 
 		//********************************
 		//********** GEOMETRY ************
